+++ lib/ui/map2/Map2HomeFilters.dart
@@ -362,13 +362,13 @@ class Map2DiningLocationsFilter extends Map2Filter {
 
   Map2DiningLocationsFilter._fromJson(Map<String, dynamic> json) :
     onlyOpened = JsonUtils.boolValue(json['onlyOpened']) ?? false,
-    paymentType = PaymentTypeImpl.fromJson(JsonUtils.stringValue(json['paymentType'])),
+    paymentType = PaymentTypeImpl.fromJsonString(JsonUtils.stringValue(json['paymentType'])),
     super._fromJson(json);
 
   @override
   Map<String, dynamic> toJson() => <String, dynamic>{
     'onlyOpened': onlyOpened,
-    'paymentType': paymentType?.toJson(),
+    'paymentType': paymentType?.toJsonString(),
     ...super.toJson(),
   };
 
@@ -403,7 +403,7 @@ class Map2DiningLocationsFilter extends Map2Filter {
       descriptionMap[searchKey] = <String>[searchText];
     }
     if (paymentType != null) {
-      String? paymentTypeValue = PaymentTypeHelper.paymentTypeToDisplayString(paymentType);
+      String? paymentTypeValue = paymentType?.displayTitle;
       if ((paymentTypeValue != null) && paymentTypeValue.isNotEmpty) {
         String paymentTypeKey = Localization().getStringEx('panel.map2.filter.payment_type.text', 'Payment Type');
         descriptionMap[paymentTypeKey] = <String>[paymentTypeValue];